Fixing delete custom and no custom computer

+++ ComputerApp/Controllers/ComputersController.cs
@@ -23,13 +23,15 @@ namespace ComputerApp.Controllers
         private readonly SignInManager<AppUser> _signInManager;
         private readonly UserManager<AppUser> _userManager;
         private readonly OrderService _orderService;
+        private readonly HelperService _helperService;
 
-        public ComputersController(ApplicationDbContext context, UserManager<AppUser> userManager, SignInManager<AppUser> signInManager, OrderService orderService)
+        public ComputersController(ApplicationDbContext context, UserManager<AppUser> userManager, SignInManager<AppUser> signInManager, OrderService orderService, HelperService helperService)
         {
             _userManager = userManager;
             _signInManager = signInManager;
             _context = context;
             _orderService = orderService;
+            _helperService = helperService;
         }
 
         // GET: Computers
@@ -51,53 +53,16 @@ namespace ComputerApp.Controllers
                 }
                 foreach (Computer item in myList)
                 {
-                    await UpdateComputerPrice(item);
-                }
-            }
-
-            dataToSendToView = LoadComputerVM(myList, ComponentList);
-            return View(dataToSendToView);
-        }
-
-        //Se actualiza el price en la tabla computer, recibe el objeto computer y actualiza el valor Price
-        public async Task UpdateComputerPrice(Computer computer)
-        {
-            double price = 0;
-            List<Component> Components = await _context.Component.ToListAsync();
-            List<ComputerComponent> ComputerComponents = await _context.ComputerComponent.Where(computerItem => computerItem.ComputerId == computer.Id).ToListAsync();
-            foreach (var computerComponent in ComputerComponents)
-            {
-                price += computerComponent.Component.Price;
-            }
-            computer.Price = price;
-        }
-
-        //Función que carga el objeto tipo ComputerVM para luego ser enviado a la vista
-        public List<ComputerVM> LoadComputerVM(List<Computer> myList, List<Component> ComponentList)
-        {
-            List<ComputerVM> dataToLoad = new List<ComputerVM>();
-            foreach (Computer item in myList)
-            {
-                ComputerVM itemComputerVM = new ComputerVM();
-                itemComputerVM.ComputerId = item.Id;
-                itemComputerVM.ImgUrl = item.ImgUrl;
-                itemComputerVM.Price = item.Price;
-                itemComputerVM.Qty = 1;
-                //itemComputerVM.TotalPrice = item.Price;
-                foreach (ComputerComponent subItem in item.ComputerComponents)
-                {
-                    foreach (var componentItem in ComponentList)
+                    if (item.ComputerComponents.Count > 0) //Solo actualizará precio cuando el computer sea Custom
                     {
-                        if (componentItem.Id == subItem.ComponentId)
-                        {
-                            itemComputerVM.Products.Add(componentItem.Name);
-                        }
+                        await _helperService.UpdateComputerPrice(item);
                     }
+
                 }
-                dataToLoad.Add(itemComputerVM);
             }
 
-            return (dataToLoad);
+            dataToSendToView = _helperService.LoadComputerVM(myList, ComponentList);
+            return View(dataToSendToView);
         }
 
         // GET: Computers/Details/5
@@ -220,9 +185,17 @@ namespace ComputerApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var computer = await _context.Computer.FindAsync(id);
-            _context.Computer.Remove(computer);
-            await _context.SaveChangesAsync();
+            string nameOfCustomComputer = "Custom Computer";
+            //var computer = await _context.Computer.FindAsync(id);
+            var computer = await _context.Computer
+                                .Where(computerItem => computerItem.Id == id)
+                                .Include(computerOrderItem => computerOrderItem.ComputerOrders).FirstOrDefaultAsync();
+            if (computer.Name == nameOfCustomComputer)
+            {
+                _context.Computer.Remove(computer);
+                await _context.SaveChangesAsync();
+            }
+
             return RedirectToAction(nameof(Index));
         }
 
@@ -269,11 +242,11 @@ namespace ComputerApp.Controllers
 
             if (orderAssociatedWUser == null)
             {
-                order.Price += GetComputerTotalPrice(dataFromView);
+                order.Price += _helperService.GetComputerTotalPrice(dataFromView);
                 order.Qty = 1;
                 order.IsCart = false;
                 order.AppUserId = myCurrentUser.Id;
-                orderId = await InsertOrderToDB(order);
+                orderId = await _helperService.InsertOrderToDB(order);
             }
             else
             {
@@ -281,12 +254,12 @@ namespace ComputerApp.Controllers
             }
 
             computer.Name = "Custom Computer";
-            computer.Price = GetComputerTotalPrice(dataFromView);
+            computer.Price = _helperService.GetComputerTotalPrice(dataFromView);
             computer.IsDesktop = true;
             computer.ImgUrl = "https://c1.neweggimages.com/NeweggImage/ProductImage/83-221-575-V09.jpg";
-            int computerId = await InsertComputerToDB(computer);
-            int computerComponentId = await InsertComponentsToComputerComponentDB(dataFromView, computerId, orderId);
-            int computerOrderId = await InsertComputerOrderToDB(orderId, computerId);
+            int computerId = await _helperService.InsertComputerToDB(computer);
+            int computerComponentId = await _helperService.InsertComponentsToComputerComponentDB(dataFromView, computerId, orderId);
+            int computerOrderId = await _helperService.InsertComputerOrderToDB(orderId, computerId);
 
             return RedirectToAction(nameof(Index));
 
@@ -334,7 +307,8 @@ namespace ComputerApp.Controllers
             {
                 return NotFound();
             }
-            computer.Price = GetComputerTotalPrice(dataFromView);
+            computer.Price = _helperService.GetComputerTotalPrice(dataFromView);
+
             idArray[0] = dataFromView.HddId; idArray[1] = dataFromView.SoftwareId; idArray[2] = dataFromView.ProcessorId; idArray[3] = dataFromView.MemoryId; idArray[4] = dataFromView.OSId;
             int i = 0;
             foreach (var item in computer.ComputerComponents)
@@ -362,73 +336,73 @@ namespace ComputerApp.Controllers
         }
 
         //-----------------------------------------END BUILD OWN COMPUTER
-        public double GetComputerTotalPrice(ComponentVM dataFromView)
-        {
-            Type type = typeof(ComponentVM);
-            int NumberOfRecords = type.GetProperties().Length;
-            int[] idArray = new int[NumberOfRecords];
-            double totalPrice = 0;
-            Component component = new Component();
-            idArray[0] = dataFromView.HddId; idArray[1] = dataFromView.SoftwareId; idArray[2] = dataFromView.ProcessorId; idArray[3] = dataFromView.MemoryId; idArray[4] = dataFromView.OSId;
-            foreach (int item in idArray)
-            {
-                component = _context.Component.Where(c => c.Id == item).FirstOrDefault<Component>();
-                totalPrice += component.Price;
-            }
-            return (totalPrice);
-        }
+        //public double GetComputerTotalPrice(ComponentVM dataFromView)
+        //{
+        //    Type type = typeof(ComponentVM);
+        //    int NumberOfRecords = type.GetProperties().Length;
+        //    int[] idArray = new int[NumberOfRecords];
+        //    double totalPrice = 0;
+        //    Component component = new Component();
+        //    idArray[0] = dataFromView.HddId; idArray[1] = dataFromView.SoftwareId; idArray[2] = dataFromView.ProcessorId; idArray[3] = dataFromView.MemoryId; idArray[4] = dataFromView.OSId;
+        //    foreach (int item in idArray)
+        //    {
+        //        component = _context.Component.Where(c => c.Id == item).FirstOrDefault<Component>();
+        //        totalPrice += component.Price;
+        //    }
+        //    return (totalPrice);
+        //}
 
         //Inserta un elemento Computer nuevo en la bd Computer y devuelve el id de este elemento
         //public async Task<int> CreateFromCode([Bind("Id,Name,Price,IsDesktop,ImgUrl,OrderId")] Computer computer)
-        public async Task<int> InsertComputerToDB(Computer computer)
-        {
-            _context.Add(computer);
-            await _context.SaveChangesAsync();
+        //public async Task<int> InsertComputerToDB(Computer computer)
+        //{
+        //    _context.Add(computer);
+        //    await _context.SaveChangesAsync();
 
-            return computer.Id;
-        }
-        public async Task<int> InsertComponentsToComputerComponentDB(ComponentVM component, int computerId, int orderId)
-        {
-            Type type = typeof(ComponentVM);
-            int NumberOfRecords = type.GetProperties().Length;
-            int[] idArray = new int[NumberOfRecords];
+        //    return computer.Id;
+        //}
+        //public async Task<int> InsertComponentsToComputerComponentDB(ComponentVM component, int computerId, int orderId)
+        //{
+        //    Type type = typeof(ComponentVM);
+        //    int NumberOfRecords = type.GetProperties().Length;
+        //    int[] idArray = new int[NumberOfRecords];
 
-            ComputerComponent computerComponent = new ComputerComponent();
-            idArray[0] = component.HddId; idArray[1] = component.SoftwareId; idArray[2] = component.ProcessorId; idArray[3] = component.MemoryId; idArray[4] = component.OSId;
-            //computerComponent.ComputerId = computerId;
-            foreach (var item in idArray)
-            {
-                computerComponent = new ComputerComponent();
-                computerComponent.ComputerId = computerId;
-                computerComponent.ComponentId = item;
-                //computerComponent.OrderId = orderId;
-                _context.Add(computerComponent);
-                await _context.SaveChangesAsync();
-            }
+        //    ComputerComponent computerComponent = new ComputerComponent();
+        //    idArray[0] = component.HddId; idArray[1] = component.SoftwareId; idArray[2] = component.ProcessorId; idArray[3] = component.MemoryId; idArray[4] = component.OSId;
+        //    //computerComponent.ComputerId = computerId;
+        //    foreach (var item in idArray)
+        //    {
+        //        computerComponent = new ComputerComponent();
+        //        computerComponent.ComputerId = computerId;
+        //        computerComponent.ComponentId = item;
+        //        //computerComponent.OrderId = orderId;
+        //        _context.Add(computerComponent);
+        //        await _context.SaveChangesAsync();
+        //    }
 
-            return computerComponent.Id;
-        }
+        //    return computerComponent.Id;
+        //}
 
-        //Inserta un elemento Order nuevo en la bd Order y devuelve el id de este elemento
-        public async Task<int> InsertOrderToDB(Order order)
-        {
-            _context.Add(order);
-            await _context.SaveChangesAsync();
+        ////Inserta un elemento Order nuevo en la bd Order y devuelve el id de este elemento
+        //public async Task<int> InsertOrderToDB(Order order)
+        //{
+        //    _context.Add(order);
+        //    await _context.SaveChangesAsync();
 
-            return order.Id;
-        }
+        //    return order.Id;
+        //}
 
-        //Inserta un elemento ComputerOrder nuevo en la bd ComputerOrder y devuelve el id de este elemento
-        public async Task<int> InsertComputerOrderToDB(int orderId, int computerId)
-        {
-            ComputerOrder computerOrder = new ComputerOrder();
-            computerOrder.OrderId = orderId;
-            computerOrder.ComputerId = computerId;
-            _context.Add(computerOrder);
-            await _context.SaveChangesAsync();
+        ////Inserta un elemento ComputerOrder nuevo en la bd ComputerOrder y devuelve el id de este elemento
+        //public async Task<int> InsertComputerOrderToDB(int orderId, int computerId)
+        //{
+        //    ComputerOrder computerOrder = new ComputerOrder();
+        //    computerOrder.OrderId = orderId;
+        //    computerOrder.ComputerId = computerId;
+        //    _context.Add(computerOrder);
+        //    await _context.SaveChangesAsync();
 
-            return computerOrder.Id;
-        }
+        //    return computerOrder.Id;
+        //}
         //public async Task<int> InsertOrderToUser(Order order)
         //{
         //    _context.Add(order);
